Split import scripts between capstone and zydis
