+++ instructionAPI/importers/capstone/aarch64/mnemonics.py
@@ -9,6 +9,6 @@ class mnemonics:
     self.dyninst_prefix = "aarch64_op"
     self.missing = None
     self.pseudo = None
-    self.capstone = _read_capstone_mnemonics(cap_dir + "/arch/AArch64/AArch64GenCSMappingInsnName.inc")
+    self.external = _read_capstone_mnemonics(cap_dir + "/arch/AArch64/AArch64GenCSMappingInsnName.inc")
     self.dyninst = _read_dyninst_mnemonics(dyn_dir + "/common/h/mnemonics/aarch64_entryIDs.h")
     self.aliases = None
